Port 25 GPU examples from C++ reference and add shader/image resources
Ported all remaining examples from TheSpydog/SDL_gpu_examples to C#:
CullMode, BasicStencil, InstancedIndexed, WindowResize, TriangleMSAA,
Clear3DSlice, DrawIndirect, TexturedQuad, TexturedAnimatedQuad,
CustomSampling, BlitMirror, GenerateMipmaps, Latency, BasicCompute,
ComputeUniforms, ComputeSampler, CopyAndReadback, CopyConsistency,
Texture2DArray, Cubemap, Blit2DArray, BlitCube, DepthSampler,
PullSpriteBatch, and ComputeSpriteBatch.

Includes HLSL shaders, image assets (BMP, HDR, DDS, ASTC), and
workarounds for SDL3-CS binding bugs in CompileComputePipelineFromSPIRV
and DownloadFromGPUBuffer.

+++ SDL-cs_gpu_examples/Examples/BasicVertexBuffer.cs
@@ -232,7 +232,7 @@ public static class BasicVertexBuffer
                 BufferSlot = 0,
                 Format = SDL.GPUVertexElementFormat.Ubyte4Norm,
                 Location = 1,
-                Offset = (uint)(sizeof(float) * 3)
+                Offset = sizeof(float) * 3
             }
         };
         var attrSize = Marshal.SizeOf<SDL.GPUVertexAttribute>();
